requirements.txt
paho-mqtt==1.6.1
bleak==0.20.2
